feat: move sshagent to separate package, update log interface
