+++ db/AccessKey.go
@@ -7,7 +7,6 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
-	"github.com/ansible-semaphore/semaphore/lib"
 	"io"
 	"math/big"
 	"os"
@@ -15,6 +14,8 @@ import (
 	"strconv"
 	"time"
 
+	"github.com/ansible-semaphore/semaphore/lib"
+	"github.com/ansible-semaphore/semaphore/pkg/ssh"
 	"github.com/ansible-semaphore/semaphore/util"
 )
 
@@ -66,14 +67,14 @@ const (
 
 type AccessKeyInstallation struct {
 	InstallationKey int64
-	SshAgent        *lib.SshAgent
+	SSHAgent        *ssh.Agent
 	Login           string
 	Password        string
 }
 
 func (key AccessKeyInstallation) Destroy() error {
-	if key.SshAgent != nil {
-		return key.SshAgent.Close()
+	if key.SSHAgent != nil {
+		return key.SSHAgent.Close()
 	}
 
 	installPath := key.GetPath()
@@ -89,11 +90,10 @@ func (key AccessKeyInstallation) GetPath() string {
 	return util.Config.TmpPath + "/access_key_" + strconv.FormatInt(key.InstallationKey, 10)
 }
 
-func (key *AccessKey) startSshAgent(logger lib.Logger) (lib.SshAgent, error) {
-
-	sshAgent := lib.SshAgent{
+func (key *AccessKey) startSSHAgent(logger lib.Logger) (ssh.Agent, error) {
+	sshAgent := ssh.Agent{
 		Logger: logger,
-		Keys: []lib.SshAgentKey{
+		Keys: []ssh.AgentKey{
 			{
 				Key:        []byte(key.SshKey.PrivateKey),
 				Passphrase: []byte(key.SshKey.Passphrase),
@@ -127,9 +127,9 @@ func (key *AccessKey) Install(usage AccessKeyRole, logger lib.Logger) (installat
 	case AccessKeyRoleGit:
 		switch key.Type {
 		case AccessKeySSH:
-			var agent lib.SshAgent
-			agent, err = key.startSshAgent(logger)
-			installation.SshAgent = &agent
+			var agent ssh.Agent
+			agent, err = key.startSSHAgent(logger)
+			installation.SSHAgent = &agent
 		}
 	case AccessKeyRoleAnsiblePasswordVault:
 		if key.Type != AccessKeyLoginPassword {
@@ -145,9 +145,9 @@ func (key *AccessKey) Install(usage AccessKeyRole, logger lib.Logger) (installat
 	case AccessKeyRoleAnsibleUser:
 		switch key.Type {
 		case AccessKeySSH:
-			var agent lib.SshAgent
-			agent, err = key.startSshAgent(logger)
-			installation.SshAgent = &agent
+			var agent ssh.Agent
+			agent, err = key.startSSHAgent(logger)
+			installation.SSHAgent = &agent
 			installation.Login = key.LoginPassword.Login
 		case AccessKeyLoginPassword:
 			installation.Login = key.LoginPassword.Login
