+++ pkg/ssh/agent.go
@@ -1,35 +1,40 @@
-package lib
+package ssh
 
 import (
+	"fmt"
 	"io"
 	"net"
 
-	"github.com/pkg/errors"
+	"github.com/ansible-semaphore/semaphore/lib"
 	"golang.org/x/crypto/ssh"
 	"golang.org/x/crypto/ssh/agent"
 )
 
-type SshAgentKey struct {
+type AgentKey struct {
 	Key        []byte
 	Passphrase []byte
 }
 
-type SshAgent struct {
-	Keys       []SshAgentKey
-	Logger     Logger
+type Agent struct {
+	Keys       []AgentKey
+	Logger     lib.Logger
 	listener   net.Listener
 	SocketFile string
 	done       chan struct{}
 }
 
-func (a *SshAgent) Listen() error {
+func NewAgent() Agent {
+	return Agent{}
+}
 
-	var err error
-
+func (a *Agent) Listen() error {
 	keyring := agent.NewKeyring()
 
 	for _, k := range a.Keys {
-		var key interface{}
+		var (
+			key interface{}
+			err error
+		)
 
 		if len(k.Passphrase) == 0 {
 			key, err = ssh.ParseRawPrivateKey(k.Key)
@@ -38,19 +43,28 @@ func (a *SshAgent) Listen() error {
 		}
 
 		if err != nil {
-			return errors.Wrap(err, "parsing private key")
+			return fmt.Errorf("parsing private key: %w", err)
 		}
-		err = keyring.Add(agent.AddedKey{PrivateKey: key})
 
-		if err != nil {
-			return err
+		if err := keyring.Add(agent.AddedKey{
+			PrivateKey: key,
+		}); err != nil {
+			return fmt.Errorf("adding private key: %w", err)
 		}
 	}
 
-	l, err := net.ListenUnix("unix", &net.UnixAddr{Net: "unix", Name: a.SocketFile})
+	l, err := net.ListenUnix(
+		"unix",
+		&net.UnixAddr{
+			Net:  "unix",
+			Name: a.SocketFile,
+		},
+	)
+
 	if err != nil {
-		return errors.Wrapf(err, "listening on socket %q", a.SocketFile)
+		return fmt.Errorf("listening on socket %q: %w", a.SocketFile, err)
 	}
+
 	l.SetUnlinkOnClose(true)
 	a.listener = l
 	a.done = make(chan struct{})
@@ -58,12 +72,13 @@ func (a *SshAgent) Listen() error {
 	go func() {
 		for {
 			conn, err := a.listener.Accept()
+
 			if err != nil {
 				select {
 				case <-a.done:
 					return
 				default:
-					a.Logger.Log("error accepting socket connection")
+					a.Logger.Logf("error accepting socket connection: %w", err)
 					return
 				}
 			}
@@ -72,15 +87,16 @@ func (a *SshAgent) Listen() error {
 				defer conn.Close()
 
 				if err := agent.ServeAgent(keyring, conn); err != nil && err != io.EOF {
-					a.Logger.Log("error serving SSH agent")
+					a.Logger.Logf("error serving SSH agent listener: %w", err)
 				}
 			}(conn)
 		}
 	}()
+
 	return nil
 }
 
-func (a *SshAgent) Close() error {
+func (a *Agent) Close() error {
 	close(a.done)
 	return a.listener.Close()
 }
