+++ services/runners/job_pool.go
@@ -1,93 +1,22 @@
-//
-// Runner's job pool. NOT SERVER!!!
-// Runner gets jobs from the server and put them to this pool.
-//
-
 package runners
 
 import (
-	"bufio"
 	"bytes"
 	"encoding/json"
 	"fmt"
-	"github.com/ansible-semaphore/semaphore/db"
+	"io"
+	"net/http"
+	"os"
+	"strconv"
+	"sync/atomic"
+	"time"
+
 	"github.com/ansible-semaphore/semaphore/db_lib"
 	"github.com/ansible-semaphore/semaphore/lib"
 	"github.com/ansible-semaphore/semaphore/services/tasks"
 	"github.com/ansible-semaphore/semaphore/util"
 	log "github.com/sirupsen/logrus"
-	"io"
-	"net/http"
-	"os"
-	"os/exec"
-	"strconv"
-	"sync/atomic"
-	"time"
 )
-
-type jobLogRecord struct {
-	taskID int
-	record LogRecord
-}
-
-type resourceLock struct {
-	lock   bool
-	holder *job
-}
-
-// job presents current job on semaphore server.
-type job struct {
-	username        string
-	incomingVersion *string
-
-	// job presents remote or local job information
-	job             *tasks.LocalJob
-	status          lib.TaskStatus
-	args            []string
-	environmentVars []string
-}
-
-type JobData struct {
-	Username        string
-	IncomingVersion *string
-	Task            db.Task        `json:"task" binding:"required"`
-	Template        db.Template    `json:"template" binding:"required"`
-	Inventory       db.Inventory   `json:"inventory" binding:"required"`
-	Repository      db.Repository  `json:"repository" binding:"required"`
-	Environment     db.Environment `json:"environment" binding:"required"`
-}
-
-type RunnerState struct {
-	CurrentJobs []JobState
-	NewJobs     []JobData            `json:"new_jobs" binding:"required"`
-	AccessKeys  map[int]db.AccessKey `json:"access_keys" binding:"required"`
-}
-
-type JobState struct {
-	ID     int            `json:"id" binding:"required"`
-	Status lib.TaskStatus `json:"status" binding:"required"`
-}
-
-type LogRecord struct {
-	Time    time.Time `json:"time" binding:"required"`
-	Message string    `json:"message" binding:"required"`
-}
-
-type RunnerProgress struct {
-	Jobs []JobProgress
-}
-
-type JobProgress struct {
-	ID         int
-	Status     lib.TaskStatus
-	LogRecords []LogRecord
-}
-
-type runningJob struct {
-	status     lib.TaskStatus
-	logRecords []LogRecord
-	job        *tasks.LocalJob
-}
 
 type JobPool struct {
 	// logger channel used to putting log records to database.
@@ -103,16 +32,6 @@ type JobPool struct {
 	config *util.RunnerConfig
 
 	processing int32
-}
-
-type RunnerRegistration struct {
-	RegistrationToken string `json:"registration_token" binding:"required"`
-	Webhook           string `json:"webhook"`
-	MaxParallelTasks  int    `db:"max_parallel_tasks" json:"max_parallel_tasks"`
-}
-
-func (p *runningJob) Log2(msg string, now time.Time) {
-	p.logRecords = append(p.logRecords, LogRecord{Time: now, Message: msg})
 }
 
 func (p *JobPool) existsInQueue(taskID int) bool {
@@ -133,42 +52,6 @@ func (p *JobPool) hasRunningJobs() bool {
 	}
 
 	return false
-}
-
-func (p *runningJob) Log(msg string) {
-	p.Log2(msg, time.Now())
-}
-
-func (p *runningJob) SetStatus(status lib.TaskStatus) {
-	if p.status == status {
-		return
-	}
-
-	p.status = status
-	p.job.SetStatus(status)
-}
-
-func (p *runningJob) LogCmd(cmd *exec.Cmd) {
-	stderr, _ := cmd.StderrPipe()
-	stdout, _ := cmd.StdoutPipe()
-
-	go p.logPipe(bufio.NewReader(stderr))
-	go p.logPipe(bufio.NewReader(stdout))
-}
-
-func (p *runningJob) logPipe(reader *bufio.Reader) {
-
-	line, err := tasks.Readln(reader)
-	for err == nil {
-		p.Log(line)
-		line, err = tasks.Readln(reader)
-	}
-
-	if err != nil && err.Error() != "EOF" {
-		//don't panic on these errors, sometimes it throws not dangerous "read |0: file already closed" error
-		util.LogWarningWithFields(err, log.Fields{"error": "Failed to read TaskRunner output"})
-	}
-
 }
 
 func (p *JobPool) Unregister() (err error) {
